Scrape the spec for more data
Let’s also scrape the spec for the list of code points for symbols that cause parse errors when they occur in raw HTML source.

Note: `invalid-character-reference-code-points.json` is identical to `invalid-raw-code-points.json` except U+000D (CR) is not included in the latter, because lone CR are converted to LF before tokenization. http://whatwg.org/html/parsing.html#preprocessing-the-input-stream